refactor: rename the color constants file and add a general message solution
