+++ frontend/src/components/FormNewCalendar/FormNewCalendar.jsx
@@ -6,7 +6,7 @@ import { Button } from 'primereact/button';
 import { RadioButton } from 'primereact/radiobutton';
 import { classNames as cn } from 'primereact/utils';
 import styles from './FormNewCalendar.module.css';
-import { Color } from '../../utils/calendarColors';
+import { Color } from '../../utils/common';
 
 export function FormNewCalendar({ setVisible, onCreateCalendar }) {
 	const defaultValues = {
